refactor: move the nested nav to a seperate component
