+++ example/src/App.tsx
@@ -4,84 +4,13 @@ import {
   useNavigationContainerRef,
 } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import { buildWebScreen, WebScreenRuleConfig } from 'react-native-web-screen';
 import { default as NativeScreen } from './NumbersScreen';
 import ErrorScreen from './ErrorScreen';
-import { createMaterialTopTabNavigator } from '@react-navigation/material-top-tabs';
 import { OnErrorCallback, Session, withSession } from 'react-native-turbo';
+import { Routes, webScreens } from 'example/src/webScreen';
+import NestedTab from 'example/src/NestedTab';
 
 const Stack = createNativeStackNavigator();
-const Tab = createMaterialTopTabNavigator();
-
-enum Routes {
-  NotFound = 'NotFound',
-  NumbersScreen = 'NumbersScreen',
-  WebviewInitial = 'WebviewInitial',
-  New = 'New',
-  SuccessScreen = 'SuccessScreen',
-  NonExistentScreen = 'NonExistentScreen',
-  SignIn = 'SignIn',
-  Fallback = 'Fallback',
-  NestedTabNative = 'NestedTabNative',
-  NestedTabWeb = 'NestedTabWeb',
-  NestedTab = 'NestedTab',
-}
-
-const webScreenConfig: WebScreenRuleConfig = {
-  baseURL: 'http://localhost:45678/',
-  routes: {
-    [Routes.WebviewInitial]: {
-      urlPattern: '',
-      title: 'React Native Web Screen',
-    },
-    [Routes.New]: {
-      urlPattern: 'new',
-      title: 'A Modal Webpage',
-      presentation: 'modal',
-    },
-    [Routes.SuccessScreen]: {
-      urlPattern: 'success',
-      title: 'It Worked!',
-      presentation: 'modal',
-    },
-    [Routes.NumbersScreen]: {
-      urlPattern: 'numbers',
-    },
-    [Routes.SignIn]: {
-      urlPattern: 'signin',
-      title: 'Sign In',
-      presentation: 'modal',
-    },
-    [Routes.NestedTab]: {
-      routes: {
-        [Routes.NestedTabWeb]: {
-          urlPattern: 'nested',
-          title: 'Nested Web',
-        },
-      },
-    },
-    [Routes.Fallback]: { urlPattern: '*', title: '' },
-  },
-};
-
-const webScreens = buildWebScreen(webScreenConfig);
-
-const NestedTab: React.FC = () => {
-  return (
-    <Tab.Navigator
-      screenOptions={{
-        tabBarActiveTintColor: '#00094a',
-      }}
-    >
-      <Tab.Screen {...webScreens.screens.NestedTabWeb} />
-      <Tab.Screen
-        name={Routes.NestedTabNative}
-        component={NativeScreen}
-        options={{ title: 'Nested Native' }}
-      />
-    </Tab.Navigator>
-  );
-};
 
 const App: React.FC = () => {
   const navigation = useNavigationContainerRef<any>();
